requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
oauthlib==3.1.1
pymongo==3.12.1
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
tweepy==4.0.1
urllib3==1.26.7
